refactor(voice): deduplicate realtime driver flow
